+++ src/xcode/ENA/ENA/Source/HTTPClient/Security/EvaluateTrust/DynamicEvaluateTrust.swift
@@ -1,0 +1,54 @@
+//
+// 🦠 Corona-Warn-App
+//
+
+import Foundation
+import ENASecurity
+
+struct DynamicEvaluateTrust: EvaluateTrust {
+
+	// MARK: - Init
+
+	init(
+		jwkSet: [Data],
+		trustEvaluation: TrustEvaluation
+	) {
+		self.jwkSet = jwkSet
+		self.trustEvaluation = trustEvaluation
+	}
+
+	// MARK: - Protocol EvaluateTrust
+
+	func evaluate(
+		challenge: URLAuthenticationChallenge,
+		trust: SecTrust,
+		completionHandler: @escaping (URLSession.AuthChallengeDisposition, URLCredential?) -> Void
+	) {
+#if DEBUG
+		// debug/review: print the chain
+		for i in 0..<SecTrustGetCertificateCount(trust) {
+			if let cert = SecTrustGetCertificateAtIndex(trust, i) {
+				Log.debug("[\(challenge.protectionSpace.host)] @ \(i): \(cert)", log: .crypto)
+			}
+		}
+#endif
+		let result = trustEvaluation.check(trust: trust, against: jwkSet)
+		switch result {
+		case .success:
+			completionHandler(.useCredential, URLCredential(trust: trust))
+		case .failure(let error):
+			Log.debug("AuthenticationChallenge failed with error \(error.localizedDescription)", log: .client)
+			completionHandler(.cancelAuthenticationChallenge, nil)
+		}
+	}
+
+	// MARK: - Public
+
+	// MARK: - Internal
+
+	// MARK: - Private
+
+	private let trustEvaluation: TrustEvaluation
+	private var jwkSet: [Data]
+
+}
