+++ src/xcode/ENA/ENA/Source/HTTPClient/Security/EvaluateTrust/_EvaluateTrust.swift
@@ -1,0 +1,7 @@
+//
+// 🦠 Corona-Warn-App
+//
+
+protocol EvaluateTrust {
+	func evaluate(challenge: URLAuthenticationChallenge, trust: SecTrust, completionHandler: @escaping (URLSession.AuthChallengeDisposition, URLCredential?) -> Void)
+}
